Update README.md
Make sure to replace '/path/to/chromedriver' with the actual path to the chromedriver executable on your system. You can download the appropriate version of chromedriver for your system from the official website (http://chromedriver.chromium.org/home).

Note that this script assumes you have the Selenium library installed (pip install selenium) and have the Chrome browser installed on your system.

The script navigates to the Entech "Contact" page, fills in the contact form with sample data (name, email, and message), submits the form, waits for a success message to appear, and then prints the success message. Finally, it closes the browser.